refactor: move store folder out of api folder
