+++ api/mapper/users.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/huy125/financial-data-web/api/dto"
-	"github.com/huy125/financial-data-web/api/store"
+	"github.com/huy125/financial-data-web/store"
 )
 
 // ToAPIUser converts a store user to an API user.
