+++ store/errors.go
